Create style.css
all the css for the first version. still gotta fix the formatting and I will work on that tomorrow but this is what i have so far. i just figured i should actualy make ocmmits on this since i havent even made a repository after working on this for like 10 hours.